Display, Screen, Component
Creation of a Display, Screen, and Component class.
The display class currently represents the whole display and allows for adding and getting screens in a tab like way.
The Screen class currently contains a lvlg page pointer, with a name.
The Component class is currently just an empty class, but it will eventually represent any UI component.