Screen, Component, Label
Implemented adding and retrieving of components from a Screen, similar to how Screens can be added and retrieved from a Display.

Added a pointer to the parent of the Component class, which will allow for future creation of different components.

Created new empty Label class, which will be implemented as a Label that displays a string of text